fix: remove unnecessary i32 cast in vision loop

+++ live2d-viewer/src/ai/vision_local.rs
@@ -79,7 +79,7 @@ pub fn infer_with_image(
         if token == eos { break; }
         tokens.push(token);
         let mut batch = LlamaBatch::new(1, 1);
-        batch.add(token, n_past + i as i32, &[0], true).ok();
+        batch.add(token, n_past + i, &[0], true).ok();
         vm.ctx.decode(&mut batch).ok();
     }
